refactor: split post helpers
